+++ Runtime/EditorComment.cs
@@ -2,10 +2,13 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using Sirenix.OdinInspector;
 using UnityEngine;
 using UnityEditor;
+using UnityEditor.Presets;
 using Object = UnityEngine.Object;
+using Random = UnityEngine.Random;
 
 namespace OcUtility
 {
@@ -17,8 +20,57 @@ namespace OcUtility
         public Color IconTint { get; }
         public Context[] Contexts;
 
+        static Dictionary<int, Preset> _playModeCopies;
+        [HideInInspector]public int guid;
+        bool _isDirty;
+        bool _enteredPlaymode;
+        [RuntimeInitializeOnLoadMethod]
+        static void Init()
+        {
+            if(_playModeCopies == null)
+                _playModeCopies = new Dictionary<int, Preset>();
+            else _playModeCopies.Clear();
+        }
+        
+        void Awake()
+        {
+            EditorApplication.playModeStateChanged += ApplyChanges;
+        }
+
+        void ApplyChanges(PlayModeStateChange change)
+        {
+            if(change != PlayModeStateChange.EnteredEditMode) return;
+            if (_playModeCopies.ContainsKey(guid))
+            {
+                var editModeInstance = Resources
+                    .FindObjectsOfTypeAll<EditorComment>().First(x => x.guid == guid);
+                _playModeCopies[guid].ApplyTo(editModeInstance);
+            }
+            EditorApplication.playModeStateChanged -= ApplyChanges;
+        }
+
+        void OnValidate()
+        {
+            if(!Application.isPlaying) return;
+            if (!_enteredPlaymode)
+            {
+                _enteredPlaymode = true;
+                return;
+            }
+            _isDirty = true;
+        }
+
+        void OnDestroy()
+        {
+            if(_isDirty)
+            {
+                _playModeCopies[guid] = new Preset(this);
+            }
+        }
+
         void Reset()
         {
+            guid = Random.Range(int.MinValue, int.MaxValue);
             Contexts = new Context[1];
             gameObject.hideFlags = HideFlags.DontSaveInBuild;
         }